dashboard/requirements.txt
Django>=3.1.7
pathlib>=1.0.1
requests>=2.25.1
django-background-tasks>=1.2.5
